[ALPS03537818] SWPM: fix type warning
[Detail] fix type warning to avoid 32 bit project build error

MTK-Commit-Id: cc246c3af46eb96cc9401378c7eafdad733b3904

CR-Id: ALPS03537818
Feature: Others

+++ drivers/misc/mediatek/base/power/swpm_v1/mt6771/mtk_swpm_platform.c
@@ -239,7 +239,8 @@ int swpm_platform_init(struct platform_device *pdev)
 		sizeof(dram_def_pwr_conf));
 
 	swpm_info("copy pwr data (size: aphy/dram = %ld/%ld) done!\n",
-		sizeof(aphy_def_pwr_tbl), sizeof(dram_def_pwr_conf));
+		(unsigned long)sizeof(aphy_def_pwr_tbl),
+		(unsigned long)sizeof(dram_def_pwr_conf));
 
 	return ret;
 }
